backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
httpx==0.27.2
gunicorn==21.2.0



